+++ backend/src/test/java/com/cafein/backend/api/CafeControllerTest.java
@@ -7,15 +7,29 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 
+import com.cafein.backend.api.cafe.controller.CafeController;
+import com.cafein.backend.domain.cafe.service.CafeService;
+import com.cafein.backend.domain.member.service.MemberService;
+import com.cafein.backend.domain.viewedcafe.service.ViewedCafeService;
+
 class CafeControllerTest extends ControllerTestSupporter {
+
+	@Mock
+	private CafeService cafeService;
+
+	@Mock
+	private ViewedCafeService viewedCafeService;
+
+	@Mock
+	private MemberService memberService;
 
 	@Test
 	void 카페_정보를_조회한다() throws Exception {
-		given(cafeService.findCafeInfoById(anyLong(), anyLong()))
-			.willReturn(CAFE_INFO_DTO);
+		given(cafeService.findCafeInfoById(any(), any())).willReturn(CAFE_INFO_DTO);
 
 		mockMvc(new CafeController(memberService, cafeService, viewedCafeService))
 			.perform(
@@ -29,16 +43,18 @@ class CafeControllerTest extends ControllerTestSupporter {
 
 	@Test
 	void 커피콩을_사용해서_카페_정보를_열람한다() throws Exception {
-		given(cafeService.findCafeInfoById(anyLong(), anyLong())).willReturn(CAFE_INFO_DTO);
+		given(cafeService.findCafeInfoById(any(), any())).willReturn(CAFE_INFO_DTO);
 
-		mockMvc.perform(post("/api/cafe/1")
+		mockMvc(new CafeController(memberService, cafeService, viewedCafeService))
+			.perform(
+				post("/api/cafe/1")
 				.contentType(MediaType.APPLICATION_JSON_VALUE)
 				.header(HttpHeaders.AUTHORIZATION, AUTHORIZATION_HEADER_ACCESS)
 			)
 			.andExpect(status().isOk())
 			.andExpect(jsonPath("$.cafeInfoProjection.name").value("5to7"));
 
-		then(memberService).should(times(1)).updateCoffeeBean(anyLong());
-		then(viewedCafeService).should(times(1)).addViewedCafe(any(), anyLong());
+		then(memberService).should(times(1)).subtractCoffeeBean(any());
+		then(viewedCafeService).should(times(1)).addViewedCafe(any(), any());
 	}
 }
