+++ backend/src/test/java/com/cafein/backend/api/OAuthLoginControllerTest.java
@@ -6,25 +6,26 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 
+import com.cafein.backend.api.login.controller.OAuthLoginController;
+import com.cafein.backend.api.login.service.OAuthLoginService;
 import com.cafein.backend.domain.member.constant.MemberType;
 
 class OAuthLoginControllerTest extends ControllerTestSupporter {
 
-	@Test
-	void 로그인화면을_호출한다() throws Exception {
-		mockMvc.perform(get("/login"))
-			.andExpect(status().isOk());
-	}
+	@Mock
+	private OAuthLoginService oAuthLoginService;
 
 	@Test
 	void 구글_토큰으로_로그인을_진행한다() throws Exception {
 		given(oAuthLoginService.oauthLogin(eq(ACCESS_TOKEN), eq(MemberType.GOOGLE)))
 			.willReturn(KAKAO_LOGIN_RESPONSE);
 
-		mockMvc.perform(post("/api/oauth/login")
+		mockMvc(new OAuthLoginController(oAuthLoginService))
+			.perform(post("/api/oauth/login")
 				.contentType(MediaType.APPLICATION_JSON_VALUE)
 				.header(HttpHeaders.AUTHORIZATION, AUTHORIZATION_HEADER_ACCESS)
 				.content(GOOGLE_MEMBER_TYPE)
@@ -39,7 +40,8 @@ class OAuthLoginControllerTest extends ControllerTestSupporter {
 		given(oAuthLoginService.oauthLogin(eq(ACCESS_TOKEN), eq(MemberType.KAKAO)))
 			.willReturn(KAKAO_LOGIN_RESPONSE);
 
-		mockMvc.perform(post("/api/oauth/login")
+		mockMvc(new OAuthLoginController(oAuthLoginService))
+			.perform(post("/api/oauth/login")
 				.contentType(MediaType.APPLICATION_JSON_VALUE)
 				.header(HttpHeaders.AUTHORIZATION, AUTHORIZATION_HEADER_ACCESS)
 				.content(KAKAO_MEMBER_TYPE)
@@ -47,31 +49,5 @@ class OAuthLoginControllerTest extends ControllerTestSupporter {
 			.andExpect(status().isOk())
 			.andExpect(jsonPath("$.grantType").value("Bearer"))
 			.andExpect(jsonPath("$.accessToken").value("access_token"));
-	}
-
-	@Test
-	void refresh_token을_이용해_access_token을_재발급한다() throws Exception {
-		given(tokenService.createAccessTokenByRefreshToken(REFRESH_TOKEN))
-			.willReturn(ACCESS_TOKEN_RESPONSE);
-
-		mockMvc.perform(post("/api/access-token/issue")
-				.contentType(MediaType.APPLICATION_JSON_VALUE)
-				.header(HttpHeaders.AUTHORIZATION, AUTHORIZATION_HEADER_REFRESH)
-			)
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.grantType").value("Bearer"))
-			.andExpect(jsonPath("$.accessToken").value("access_token"));
-	}
-
-	@Test
-	void 로그아웃을_진행한다() throws Exception {
-		final MvcResult result = mockMvc.perform(post("/api/logout")
-				.contentType(MediaType.APPLICATION_JSON_VALUE)
-				.header(HttpHeaders.AUTHORIZATION, AUTHORIZATION_HEADER_ACCESS)
-			)
-			.andExpect(status().isOk())
-			.andReturn();
-
-		assertThat(result.getResponse().getContentAsString()).isEqualTo("logout success");
 	}
 }
